+++ NEWS.md
@@ -1,5 +1,7 @@
 # dbplyr 1.1.0.9000
 
+* Improvements to `sql_optimise()` (#3062) (@edgararuiz)
+
 * Adding support for stringr functions: `str_length()`, `str_to_upper()`, `str_to_lower()`,
   str_replace_all()`, `str_detect()`, `str_trim()`. Regular expression support varies 
   from database to database, but most simple regular expressions should be ok (@edgararuiz)